serialize errors consistently
