+++ packages/kit/src/runtime/server/endpoint.js
@@ -1,6 +1,6 @@
 import { to_headers } from '../../utils/http.js';
 import { hash } from '../hash.js';
-import { is_pojo, normalize_request_method } from './utils.js';
+import { is_pojo, normalize_request_method, serialize_error } from './utils.js';
 
 /** @param {string} body */
 function error(body) {
@@ -134,18 +134,4 @@ export async function render_endpoint(event, mod, options) {
 			headers
 		}
 	);
-}
-
-/**
- * @param {Error} error
- * @param {import('types').SSROptions} options
- */
-function serialize_error(error, options) {
-	const object = {
-		name: error.name,
-		message: error.message,
-		stack: options.get_stack(error)
-	};
-
-	return JSON.stringify(object);
 }
