+++ packages/kit/test/apps/basics/src/routes/errors/page-endpoint/__error.svelte
@@ -5,6 +5,7 @@
 <pre>{JSON.stringify(
 		{
 			status: $page.status,
+			name: $page.error.name,
 			message: $page.error.message,
 			stack: $page.error.stack
 		},
